Updated usage of refreshApiPods
Moved refreshApiPods to a separate hook file.

Added refreshApiPods to email-settings.patch:after hooks.
The email client is configured on startup and
does not change if appConfig changes.
